feat: Open a shared Notes
The stack does all the hard work to return us a few informations that
we have to build to create the right URL to open a notes. It works
even if the notes is shared from link or cozy to cozy.

The stack also handle the fact that the Notes app is not installed and
display a page with a button to redirect to the store. So no need to
do the work from the front end.