feat: Add selector to get the content of a folder from redux
We need in actions to know if a file is present in a folder. For this,
we need to be able to extract from cozy-client's queries slice the
folder content.

We use the name of the queries to find queries that are for a particular
folder id and we return the query which is the most recent one