feat: We can have several actions with the same name
In order to display within the menu a few separators, we created an
`hr` action that display a separator. Since we don't want this separator
in the SelectionBar, we remove them before passing the props to
SelectionBar.

refactor: Extract ActionsItems in a separate file to facilitate upcoming
refacto to use same items than the MoreMenu.