feat: Redirect to trash if the folder is trashed
Previously, if you manually go to /folder/trashedFolder, it would
display the Drive view with all actions on the folder / files. But we
don't want that. So we redirect to the TrashView.

Also, some apps (like Notes) are redirecting to the note's folder in
Drive without checking if the note is trashed or not.

This is a Drive responsibility to redirect at the right place.