fix: ScrollToTop on iOS
The keyboard is opened since we have a autofocus to true on an input.
But we have issue on iOS with keyboard since the viewport is not resized

So to be sure we're displaying the right thing, we scrollToTop in order
to not displaying the blank added by iOS bellow the content.