fix: Missing last execution timestamp
We used to store a timestamp of the last execution of the service,
which is useful to know if the service should be run, depending on the 'running' and 'postponed' status.
However, this timestamp was not always set, leading to missing
information.
Therefore, we prefer to remove this timestamp and now use the
`cozyMetadata.udpatedAt` field, that we are sure is always set, as the
photo settings are updated at the beggining of the execution.