fix: Call the hooks in the right order
If not, we got undefined for currentIndex when opening
the viewer. Let's set the index before opening it.

Tried to add a test on the component, but I can't
handle it since we've a react-test-lib test but
file-opener is done with HammerJS. Click will not
fire the right event... We should test our
hooks call in order to test the component