refactor: Extract useSharingInfos from SharingBanner and PublicToolbar
By taking SharingBanner out of PublicToolbar and making them both standalone, we had to put the useSharingInfos hook in both of them. This poses a problem of query optimization if we use both components in the same parent (the hook being executed twice). The idea here is to put the hook on the parent.