feat: Replace the span by a link for onlyoffice document
I used a simple <a> instead of <Link> of react-router, because this one must have a route in the format '/myroute' rather than '/#/myroute' contrary to what replaceCurrentUrl() of createFileOpeningHandler expects. And I preferred to export this route in a createOnlyOfficeFileRoute() method to mutualize.

Also, with <Link> it would have been necessary to make a reference like this `innerRef={el => this.myRef = el}` because <Link> expects to have an innerRef of type 'func' or 'string', and thus to change the reference on the span as well, as well as on the instantiation of Hammer.