feat: Add rerouting for shared document from cozy to cozy
We remove condition to onlyoffice private route. In the case of a cozy to cozy share, we need to go to the onlyoffice route because it is only here that the query to the stack is made to know if the document is shared or not. And in this case a redirection to the host instance is done if necessary.