fix: Pass the right arguments to the cozy-bar
We decided to use data-cozy={{}} to get access to
the variables injected by the cozy-stack.

By removing data-cozy-*, cozy-bar was lost since
the bar will look into cozy-data-appslug to detect
the slug of the app. Since the bar can't do that
anymore, we need to pass all the args during the
init phase.