fix: Don't display MoreMenu if there is no file
Previously Create buttons were rendered inside the MoreMenu.
Having a condition on hasWriteAccess made sense at this time.

Now that Create buttons are in the Fab / dedicated button
it doesn't make sense anymore to check for this right. We only
display the Selection button in this MoreMenu. So testing the
number of files is only the right check.