feat: Download encrypted files in web viewer
When opening an encrypted file, the fallback viewer appears, allowing to download the file. We now transparently decrypt the file so the user can get its decrypted file on disk.
Note this does not support mobile download nor the top-right download button, as this one is directly handled by cozy-ui.