feat: Do not use client and vaultClient from store anymore
Thanks to redux-thunk, it is possible to access client and vaultClient
instances directly from the store, which is useful in non-react
contexts.
However, this is not satisfactory as we rely too much on the store and
wish to have standalone components.
Thus, we remove this from the upload module and pass clients as function
params instead.

chore: Upgrade cozy-keys-lib

fix: Tests