fix: Init vault through provider
We used to init the vault by manually creating a WebVaultClient
instance. However, this requires a URI, which is not set in mobile
context, when the user is not onboarded. This was causing a crash on the
first launch.
We fix this by using the VaultProvider, which is now capable to
automatically set the URI thanks to the given cozy-client instance.