feat: Use FileImageLoader from Cozy-UI
ImageLoader has been merged in Cozy-UI since
a few month. We can now use it from UI.

Also, we use the new `tiny` format to display
the thumbnails in the FileList for Drive since
tiny is for 96x96. We expect that a folder
that contains a lot of photos will be displayed
faster.