fix: Use replace instead of push on init router
Currently, when going to drive from cozy-homepage,
the user gets stuck in an endless loop if trying to navigate backward.
This is because the app router is instantiated on a single hash,
then push the pathname after it.
An easy way to fix that is to use a router.replace,
so that way the initial hash route is deleted from the stack.
Now the user can navigate backward without any issue.