fix: Shared OO document displayed a blank page
When there was a cozy to cozy sharing from Alice to Bob with a OO file, and Bob share the OO file with a public link, the public link displayed a blank page. The code to open the file on an other instance was bypassed when the file was public.