chore: Increase bundlemon service size
I updated cozy-client, which led to a size increase in the service size. The previous limit was arbitrary. The new limit is set to 500kB because it is the max size we should have for a service JS file.

I also updated the maxPercentIncrease to 0.5% because 0.1% for small files was not enough.