Use the name instead of the label or icon for the action key generation
The issue with the use of icon and label as value for the key is that
those fields can change based on the device used. Moreover the label
is at the moment a work used to fetch the correct string for I18n but
by using this technique we need to search for the string for each
component.