fix: Upload mostly not working on root dir
Previsouly we were not querying anymore the current folder
by making a getById() when we arrive on the /folder route.

The result was that we didn't have the document
`io.cozy.files.root-dir` inside the CozyClient store, but
by calling `getDisplayedFolder()` we called indirectly
`getDocumentFromState` and since we didn't have any reference
to the folder inside the store, we were returning null.

Since we returned null, upload was not working.

This commit fix the issue by:
- adding a query to fetch the current folder inside the Drive
View
- not relying anymore directly on redux. Let's use the
Cozy Client's magic.