fix: Navigation from Trash to Recent was displaying a blank page
Since we started to call useQuery() instead of reading the redux
store 162aa9a8611165e094a82f447bddbcd707fdc500

We were able to call useQuery without a defined currentFolderId

Since we can't call getById() with an undefined ID ATM even
with enabled to false, we need to handle that in Drive's code.

How currentFolderId can be undefined between the change of
trash to recent view? Because we have a lot of re-render
before moving from one view to an other. Not relying on
redux is part of this effort to remove all this re-render
it's a long work.