fix: Dedup of date-fns
Since the cozy-client upgrade to 34.7.1 date-fns is a needed
dep for a few of its models. But we don't need them in Drive
for now and we have a few issues with the tree shaking of
cozy-client.

(see https://github.com/cozy/cozy-client/pull/1274)

So the workaround is to force the resolution to date-fns 1
which is required by drive, in order to not duplicate the
package.

We should not have issue with that. When we'll upgrade
date-fns to v2 in drive & ui, then, we'll be able to
remove this line.