feat: Refresh page after onlyoffice idle session
OnlyOffice displays an alert if the user stays too long on a document without acting. This alert asks him to reload the page. We capture the event and automatically refresh the page. After the refresh, the page is back in reading mode so the user can click to edit it again