fix: Revert refresh after OnlyOffice idle session
This update changes the OnlyOffice key on the stack side without updating the front side. This side effect leads to an outdated version message on OnlyOffice, which is not desirable from a user experience point of view. We have chosen to go back until we get a better user experience.

Refs: 8276fed5b96d1c655fac23d107deee0d755c147d