fix: Show rename photo album input on desktop mode
The problem stems from a conflict with the MUI Menu. The menu goes into focus when it is opened for accessibility reasons. This causes the input to lose focus and disappear just afterwards because it blurs. I've disabled autoFocus for this specific case until I find a fix in cozy-ui.