fix(Toolbar): Render for each properties
Since migrating to react-router v6, the component is not rendered every time, which creates inconsistencies in the sharing modal.  This is because we select the properties that would trigger rendering and the new properties from the react-router migration were not selected. I've decided to remove this check to avoid similar errors in the future. I took the opportunity to convert the component to functional