test(Public): Remove the act error caused by an useEffect leak
The useEffect inside usePublicWritePermissions is leaking only during tests. The problem seems to come from fetchOwn from cozy-client because it use fetch and we don't clean it. We didn’t find a way to fix it so we decide to mock the problematic hook. To solve it would probably need to add an abortController to manage fetch into cozy-client

Co-authored-by: zatteo <zatteo@users.noreply.github.com>