refactor(actions): Migrate share to ActionsMenuItem
Actions are interpreted in two places, in ActionsMenu and in  SelectionBar. Before migration, action function always received an array of documents in both case. After migration, ActionsMenu now sends the document directly, while the SelectionBar sends an array of documents. To manage the two cases, I used an verification on if the params was an Array or not. In the case, we expected a single file if its an array I take the first element otherwise I put the single file into an array.