fix: Add terser plugin
There is a conflict on the terser plugin since we added
storybook to harvest it was impossible to build the drive
app.

In order to avoid this issue, I did same process as Crash on home. I ran
a yarn why terser-webpack-plugin

And install the same version as the one from cozy-scripts