feat: Add tooling around service development
To develop new service, we use cozy-jobs-cli. To make it easier to use, given that we have 2 apps in the same repository, I've created a specific command for each one. For example, you can start the onPhotoTrashed service with yarn service:photos build/photos/services/onPhotoTrashed/photos.js. Be careful to use the build file