fix: OnlyOffice back button redirect when coming from public folder
The redirection of the back button requires adaptation when opening an onlyoffice file from a folder shared by a link or the preview of a shared folder. You need to stay on the same cozy and share the pathname and sharecode to remain authenticated during browsing. However, this behaviour must be different for public files accessed directly, where you must return to the instance from which you came. This commit allows the first behaviour to be taken into account without breaking the second.