fix: Add harvest route
Since the addition of the HarvestBanner we can configure a
konnector from the banner but in order to do that, we need
to add the harvest route component.

It's been a bit of a struggle, because of the Routes. But
Celestin found the answer:

When we `useLocation` from a component, if this component
is used on the parent route, then its navigation will be
based of this route, not the closest.

Before, if we did: navigate('./harvest') from the DriveView
component, even if the route matched /folder/:folderId,
the result was: /folder/harvest and not /folder/:folderId/harvest

This is why we changed the route definition.

We also remove the need of /folder/file/:fileid since this is not needed
anymore because now even if the folder is not defined (aka root dir)
we redirect to /folder/root.dir/file/:fileid

Since we import Harvestroutes, we need to install leaflet for now...