fix: Don't rely on the createdByApp folder metadata
But use the first doc instead. Because if the folder is created by the
home, then the createdByApp will be `home` instead of the konnector slug

So I'm reading the info from the first file of the folder to get the
information.

This is performant, but I don't think it's the correct request. We
should query the triggers & get the destination folder instead.