test: Declare a store after client initialization with createMockClient
The aim of this PR is to be able to use createMockClient with a custom store. To work, createMockClient forces the creation of a store in CozyClient. During the setup of Drive store we try to apply this new store to CozyClient. However, CozyClient allow to set it once. This commit adds some exceptions to make the two work together during unit test.