fix: Make sure that the file has a path after the real-time update
Updating files in real time only returns the couchDB document, which has no path for files, only for folders. This commit adds a mechanism to make sure you have one before adding it to the cozy-client store. A file without a path creates problems during sharing checks, for example when moving a file.