fix: Can open shortcut even on preview page
We had a bug that caused an issue where if we wanted
to open a shortcut from the preview page, it wouldn't
work.

Indeed, since we were only testing the public case
and forgetting the preview case, it couldn't work.

So now, we rely on the props to be able to know if we
are public or not instead of just checking the url.