fix: Recent view query was malformed
After 6e1041f2c21c6dbc0802464af3c57f30ffedb4a9, the recent view was
broken, because it introduced the `dir_id` sort before the `updated_at`
sort, which is problematic, as the `dir_id` is meaningless for the user.
Thus, we fix this by using a simple updated_at sort and rewrite the
query to use a partial index for better performances.