feat: Open note inside public repository
The opening of the notes is subtle. The note is always opened in the instance of the note creator. This is especially true for shared notes or notes in shared folders. In this case, you need to contact the cozy-stack to find out where to open it. As this can take a little time, we use a redirection page rather than pre-calculating all the links when a folder is displayed