fix(DeleteItem): Missing props
The `onLeave` & `isSharedWithMe` props have not
been transmitted for a long time.
`onLeave` is no longer useful, it used to call a stack
route to revoke the user, but the approach has since
changed and deleting an io.cozy.files revokes the user.

`isSharedWithMe` is still useful, allowing you to
customize the wording.