fix: Declare `selected` variable before accessing it on `File.jsx`
With the Rsbuild migration, the transpiled code now use `let` and
`const` features contrary to the previous Webpack configuration that
converted them to `var`

This highlight some bug that we previously missed where some variables
were used before being declared. This did not produce any throw when
using `var` due to hoisting, but this may produce bug as the accessed
variable may be undefined

Warning: This would fix the `fil-content-row-selected` class
attribution, however this class is not compatible yet with dark theme,
so we want to enforce `undefined` until we fix the theme. Until then
the behaviour will be like before when the `selected` variable was
always `undefined`