fix: Delete confirmation icon and spacing
Icon was not displayed in a Cozy to Cozy share (it was the only
deletion text without icon) because we tried to pass to Icon
component an icon that does not exist. And when the deletion text
was too long, the icon was shrink to a very small icon.

I also adjusted spacing to have better result.