feat: Disable search on bar on mobile
cozy-bar now embeds its own search for desktop and mobile. But we do
not want to use the mobile search of cozy-bar for now because it
relies on dataproxy which is not ready on mobile.

So we disable the search of cozy-bar in mobile. We still have the old
search of drive.