test: Fix `selectors` unit tests
Since cozy/cozy-client#1594 the `selectors.spec.js` test where failing

This was due to incorrect mock for `client.requestQuery` that would
return incorrect values with mixed folders and files where
queryDefinition did request for only one type

With previous cozy-client implementation, the `extractAndMergeDocument`
would always return a new reference for the documents list. Then on the
store's Reducer, the `haveDocumentsChanged` variable would always be
`true`, so the queries helpers would try to auto update the queries and
so would replay them with Sift

The combination of incorrect mock + incorrect `dir_id` fields would
result to Sift not being able to retrieve existing files and so it
would chose to empty the entire `data` field from the `file` query in
the redux store

Also the `directory` query would contains both files and folders, so
the test did see 13 items in queries, but for the wrong reason:
- Instead of having: `files query=> 10 files` and `directories query => 3 folders`
- We had: `files query=> 0 files` and `directories query => 3 folders and 10 files`

With the new cozy-client implementation, the `extractAndMergeDocument`
would return the reference for the documents list as both `file` and
`directory` queries where mocked to receive the exact same content, so
the documents list was identical. Then the on the store's Reducer, the
`haveDocumentsChanged` variable would be `false`, and so the queries
helpers would not replay queries with Sift

So we obtained the following result:
`files query=> 3 folders and 10 files` and
`directories query => 3 folders and 10 files`, resulting to 26 items

This shows that with the new implementation, cozy-client is more
sensitive to impossible tests configurations

By fixing the mock, we now correctly have: `files query=> 10 files` and
`directories query => 3 folders`

The fix on the `dir_id` is not mandatory here, because there are no
intersection between both queries results, but we chose to still add it
in order to have a clean test setup