fix: Empty toolbar in OnlyOffice in public
In ed941cd1bffb09f0cd075c34fdc77162453219f4, we replaced the query to
get the file in OnlyOffice toolbar to avoid performance issues with
path calculation.

But the new query return a 403 in public context due to stack
limitation (files/_find with where clause vs files/:id:).
So here I switch between queries depending on public context.

We won't be able to display path in toolbar in OnlyOffice in public,
but :
- it is currently not displayed neither
- it is useless when sharing a direct file (it would make sense when
sharing a folder)