fix: Favorith path not working on mobile
To support routing in tabs that are not based on a folder (for example /sharings, /recent, /favorites), we used to duplicate every routes so that we can browse files and folders in /sharings or /recent.

Here, we take advantage of /favorites routes to be broken to try another approach. When in the /favorites tab, we set every link to default folders (folder/:id) and files (folder/:id/file/:id) links. So we do not browse files and folders in /favorites/xxx route anymore.

When clicking on a file or a folder :
- browser back button goes back to /favorites
- Drive navigation back button goes to folder higher in folder hierarchy