feat: Add basic Docs support
Add support to Docs notes: https://github.com/suitenumerique/docs

This allows to:
- Create a Docs note, that will make a redirection to the Docs app
- Display the icon of a Docs note when listed in a directory

A Docs note is represented as a io.cozy.files, with a special extension,
`.docs-note`, and its Docs id, stored in `metadata.externalId`.
The mimeType is `text/markdown` as it is indeed a markdown file,
exported from Docs, but the extension is not `.md` to avoid any local
edition in an external editor, that would be eventually overwritten
by Docs.

A flag `docs.embedded-app-url` is required to enable this feature, which
is the same flag used by the wrapper app cozy-lasuite-docs