feat: Improve performances of actions causing multiple realtime updates
When receiving a batch of realtime events, we used to treat them
sequentially and do the following:
- Query the parent folder if it's a file to compute the path
- Dispatch the mutation

This was highly inefficient, because the dispatch is a costful
operation, especially if the store is filled with documents and queries.
Indeed, each dispatch will trigger the full reevaluation of the existing
queries against the documents in the store.

So, to improve this behaviour we now introduce a debounce mechanism for
relatime events: we wait a fixed amount of time before processing the
events all at once.
To do this, each new realtime event is temporarily stored in a buffer,
and removed once treated.

This impact all treatments involving multiple realtime events, such as
trash, restore, or deletion from trash.

Note the debounce does not negatively affect single realtime events, as
we force the first event to be treated immediatly (and debounce only the
next ones if they occur less than 500ms after).

To show how impactful this is, we measured how long takes a folder to be
put in the trash. The folder contains 1K files.

- Before : ~61000 ms
- After : ~700 ms