feat: Mutualize queries by id in store
We had 3 different queries to do the same thing: query a io.cozy.files
by its id.
Typically for a directory `abc`, we had 3 queries named:
- `io.cozy.files/abc`
- `onlyfolder-abc`
- `folder-abc`

The only specialization was for the `folder-abc` that was including
`parent` and `encryption`, but it does not seem used anymore.

This is useless and hurt performances: as we duplicate such queries,
they are kept in the store and reevaluated independently even though the
id and the results are the same. It also takes memory for nothing.
Thus, let's mutualize.