feat: Upgrade packages to improve sharings retrieval perfs
When retrieving the sharings, we used to get *all* the shared documents
as well. After investigation, this is not useful, as we typically rely
on the shared directories to determine if a path is in a shared directory.
Thus, this response was useless and hurt performances.

Retrieving sharings with 544 docs takes ~500ms while it takes ~50ms
without the docs