feat: Add support for offline navigation in web
We now support offline navigation thanks to the `DataProxyLink`.  By
default, all request go through the `StackLink`, but it now detects if
the user is offline, and then forward data queries to the DataProxy,
which will run them on PouchDB.

Note this only work for read queries that rely on `io.cozy.files`: we
typically do not support offline download, or write operations such as
renaming yet.