fix: Broken public sharing pages
The public pages were broken since the upgrade of the
`cozy-dataproxy-lib`:
https://github.com/cozy/cozy-drive/commit/2107f71266077ed10ef1671e4c7084ae1e61ae48
This is because the `DataProxyProvider` renders the children only once
the DataProxy is correctly initialized. See
https://github.com/cozy/cozy-libs/commit/bdb16705103fa222d631750403d951eee61f4868

In the case of public sharing, the initialization fails because we try
to get an intent, which redirects to the sso as the recipient is not
logged in. And this redirection is blocked by the CSP.

So, to avoid this, we simply do not use the `DataProxyProvider`
when we are in a public sharing, as it is not useful for now.

Nevertheless, this DataProxy init behaviour seems error prone and
difficult to debug, so, we'll try to simplify it to avoid potential
future issues in apps.