feat: Display menu on sidebar shared drive with manage access action
A icon can now appear on hovering a shared drive item in sidebar.
When clicking this icon, it displays an action menu. This action menu
for now allows to open a manage access modal.

Opening this manage access modal is done with a redirection. It could
be improved (so without redirection) later.