fix: Impossible to create folder on Chrome Android
useShiftArrowsSelection needs to focus the first row of a table in
order to work when nothing is selected.

But this focus make it impossible to create folder on Chrome Android.
Without understanding deeply why, Chrome Android (and only Chrome
Android) make FilenameInput lose focus when this hook is used.

So here I decided enable this hook only in desktop because it makes
no sense in mobile so it will avoid doing useless work in mobile and
solve the issue in the same time.